restore drop argument changed/inverted (plus more)
mongorestore feature to not drop by default was --restoreDrops but is
now --noDrop
Also the suggestion to change [[ -d current]] to "sudo test -d current"
because of permission issue (for a rare few) seemed better.
There was also an issue with getting making the local directory if it
doesn't exist (and is not a file).